+++ scylla_api/rest/scylla_rest_client.py
